Extracted AutoRotator from Character
