+++ ShanghaiBloodSports/Assets/AutoRotator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b68319617938e0d41a0f1ead7cf92a04
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
