+++ ShanghaiBloodSports/Assets/Prefabs/Character.prefab
@@ -744,6 +744,7 @@ GameObject:
   - component: {fileID: 2609417324641571754}
   - component: {fileID: 7931033671868165058}
   - component: {fileID: 7257004311886182847}
+  - component: {fileID: 663120146686638618}
   m_Layer: 8
   m_Name: Character
   m_TagString: Untagged
@@ -864,6 +865,18 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
+--- !u!114 &663120146686638618
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2609417324641571755}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b68319617938e0d41a0f1ead7cf92a04, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &2609417324920767474
 GameObject:
   m_ObjectHideFlags: 0
