+++ ShanghaiBloodSports/Assets/Scripts/Character.cs
@@ -8,8 +8,8 @@ public class Character : MonoBehaviour
     private TextMesh healthIndicator;
     private double health = 100;
     private double guard = 100;
-    private static readonly int FACE_RIGHT = 180;
-    private static readonly int FACE_LEFT  = 0;
+
+    public Character Opponent { get; private set; }
 
     public double Health {
         get => health;
@@ -23,30 +23,12 @@ public class Character : MonoBehaviour
     void Start()
     {
         healthIndicator = transform.GetChild(0).GetComponent<TextMesh>();
-
+        Opponent = FindObjectsOfType<Character>().Where(c => c != this).Single();
     }
 
     // Update is called once per frame
     void Update()
     {
-        if(ToLeftOfOpponent())
-        {
-            transform.eulerAngles = new Vector3(transform.rotation.x, FACE_LEFT);
-        }
-        else
-        {
-            transform.eulerAngles = new Vector3(transform.rotation.x, FACE_RIGHT);
-        }
-    }
 
-    public Character Opponent => FindObjectsOfType<Character>().Where(c => c != this).Single();
-
-    private bool ToLeftOfOpponent()
-    {
-        var opponent = Opponent;
-
-        var relativeXPosition = opponent.transform.position.x - transform.position.x;
-
-        return relativeXPosition > 0;
     }
 }
